requirements.txt
openpyxl
xlrd==1.2.0
msedge-selenium-tools
selenium